+++ Test/WhiskeyTestTasks.psm1
@@ -583,3 +583,23 @@ function GenerateOutpuTaskWithCmdletBinding
 
     $Output
 }
+
+function RunsCommandThatSetsLastExitCodeTo1
+{
+    [CmdletBinding()]
+    [Whiskey.Task('RunsCommandThatSetsLastExitCodeTo1')]
+    param(
+
+    )
+
+    # Have to run powershell via Invoke-Expression otherwise when run by our Test-Whiskey.ps1 script, the commands fails
+    # with an error that '-NonInteractive' is not a command.
+    if ($PSVersionTable['Edition'] -eq 'Core')
+    {
+        Invoke-Expression 'pwsh -NoProfile -NonInteractive -Command ''exit 1'''
+    }
+    else
+    {
+        Invoke-Expression 'powershell -NoProfile -NonInteractive -Command ''exit 1'''
+    }
+}